Modifies the dsl.kt file to be MultiPlatform
This required some adjustment to other classes, which were also brought to MultiPlatform, such as AssertionErrorCollector and modifications to the Failures object.
