+++ kotlintest-assertions/src/commonMain/kotlin/io/kotlintest/Failures.kt
@@ -26,10 +26,20 @@ expect object Failures {
   /**
    * Central point for error creation
    *
-   * This method is centralizes the creation of Assertion Errors on all platforms.
+   * This method centralizes the creation of Assertion Errors on all platforms.
    * The errors might be tempered with depending on the platform and its features. These changes may include StackTrace
    * manipulation, for example.
    *
    */
   fun failure(message: String, cause: Throwable? = null): AssertionError
+  
+  /**
+   * Returns [throwable] after KotlinTest modifications
+   *
+   * This method, along with [failure] centralizes the creation of Assertion Errors on all platforms.
+   * The errors might be tempered with depending on the platform and its features. These changes may include StackTrace
+   * manipulation, for example.
+   *
+   */
+  fun <T: Throwable> modifyThrowable(throwable: T): T
 }
