+++ kotlintest-assertions/src/commonMain/kotlin/io/kotlintest/matchers/AssertionErrorCollector.kt
@@ -1,6 +1,7 @@
 package io.kotlintest.matchers
 
 expect object AssertionErrorCollector {
+  @PublishedApi internal var shouldCollectErrors: Boolean
   fun collectOrThrow(error: Throwable)
   fun throwCollectedErrors()
   fun setClueContext(context: String?)
