--- kotlintest-assertions/src/commonMain/kotlin/io/kotlintest/should.kt
@@ -1,4 +1,0 @@
-package io.kotlintest
-
-expect infix fun <T> T.should(matcher: Matcher<T>)
-expect infix fun <T> T.shouldNot(matcher: Matcher<T>)
