+++ kotlintest-assertions/src/jsMain/kotlin/io/kotlintest/dsl.kt
@@ -1,0 +1,10 @@
+package io.kotlintest
+
+
+actual fun diffLargeString(expected: String, actual: String, minSizeForDiff: Int): Pair<String, String> {
+  TODO()
+}
+
+internal actual fun createEqualsError(message: String, expectedRepr: String, actualRepr: String): Throwable {
+  return AssertionError(message)
+}
