--- kotlintest-assertions/src/jsMain/kotlin/io/kotlintest/should.kt
@@ -1,9 +1,0 @@
-package io.kotlintest
-
-actual infix fun <T> T.should(matcher: Matcher<T>) {
-  TODO()
-}
-
-actual infix fun <T> T.shouldNot(matcher: Matcher<T>) {
-  TODO()
-}
