+++ kotlintest-assertions/src/jvmMain/kotlin/io/kotlintest/Failures.kt
@@ -33,12 +33,26 @@ actual object Failures {
    * StackTrace only.
    */
   actual fun failure(message: String, cause: Throwable?): AssertionError = AssertionError(message).apply {
-    if (shouldRemoveKotlintestElementsFromStacktrace) {
-      removeKotlintestElementsFromStacktrace(this)
-    }
+    
+    modifyThrowable(this)
     initCause(cause)
   }
-
+  
+  /**
+   * Returns [throwable] after KotlinTest modifications
+   *
+   * This method, along with [failure] centralizes the creation of Assertion Errors on all platforms.
+   * The errors might be tempered with depending on the platform and its features. These changes may include StackTrace
+   * manipulation, for example.
+   *
+   */
+  actual fun <T : Throwable> modifyThrowable(throwable: T): T {
+    if (shouldRemoveKotlintestElementsFromStacktrace) {
+      removeKotlintestElementsFromStacktrace(throwable)
+    }
+    return throwable
+  }
+  
   /**
    * Remove KotlinTest-related elements from the top of [throwable]'s stack trace.
    *
